requirements.txt
Django>=3.2,<4.0
channels>=3.0.0
daphne>=3.0.0
djangorestframework>=3.12.0
djangorestframework-simplejwt>=5.0.0
django-oauth-toolkit>=1.5.0
social-auth-app-django>=4.0.0
Faker>=8.0.0
coverage>=5.5
pytest>=6.2.0
pytest-django>=4.4.0
pytest-asyncio>=0.15.0
